test(cli): add Mermaid/DOT/ASCII renderer stdout snapshots
Extracts the existing CLI runner plumbing out of DependencyGraphCLITests.swift
into CLIRunner.swift (internal, shared within the test target) so a new
GraphRendererSnapshotTests suite can reuse the exact invocation pipeline.

Adds three @Test methods that run the CLI against a minimal two-container
fixture (one root container wiring an APIClient singleton + one feature
container consuming it as .input) and snapshot the stdout verbatim for
Mermaid, DOT, and ASCII. The fixture is intentionally tiny so snapshot
diffs stay readable when the renderers change.

Infrastructure:
- Package.swift: InnoDIDependencyGraphCLITests now depends on
  InnoDITestSupport so it can call assertTextSnapshot; __Snapshots__
  excluded from the target.
- .gitattributes: mark CLI snapshot .txt files as linguist-generated and
  pin eol=lf, matching the macro snapshot policy.
- Tools/record-cli-snapshots.sh: convenience wrapper around
  `INNODI_RECORD_SNAPSHOTS=1 swift test --filter …GraphRendererSnapshotTests`.

Existing substring-based assertions in DependencyGraphCLITests.swift are
kept as complementary coverage; the snapshots catch any other regression.
